requirements.txt
# ===== FastAPI 서버 =====
fastapi==0.115.0
uvicorn==0.30.6

# ===== 설정 / 환경변수 =====
pydantic==2.12.5
pydantic-settings==2.4.0
python-dotenv==1.0.1

# ===== HTTP Client =====
httpx==0.27.2

# ===== MongoDB (Chat Log) =====
pymongo==4.7.2

# ===== Vector DB (Qdrant) =====
qdrant-client==1.9.1
sentence-transformers==3.0.1

# ===== Kafka =====
kafka-python==2.0.2